General update

+++ src/uai.h
@@ -3,6 +3,8 @@
 
 #include <cinttypes>
 
+#include "moves.h"
+
 #define DEF_DEPTH 5
 #define MAX_DEPTH 50
 
@@ -39,6 +41,6 @@ extern Settings settings;
 void uai();
 
 void infoString(const int depth, const int score, const uint64_t nodes,
-                const int duration, std::vector<Move> pv);
+                const double duration, std::vector<Move> pv);
 
 #endif // #ifndef UAI_H_
